third_party/WebKit/LayoutTests/fragmentation/become-unfragmented-with-lines.html

<!DOCTYPE html>
<p>The word "PASS" should be seen 7 times below.</p>
<div id="container" style="position:relative; overflow:hidden; line-height:20px;">
    <div id="multicol" style="float:left; columns:3; column-fill:auto; height:70px; orphans:1; widows:1; text-align:right;">
        <div style="width:6em;">
            <div id="inner">
                PA<br>
                PA<br>
                PA<br>
                PA<br>
                PA<br>
                PA<br>
                PA<br>
            </div>
        </div>
    </div>
    <div style="float:left;">
        SS<br>
        SS<br>
        SS<br>
        SS<br>
        SS<br>
        SS<br>
        SS<br>
    </div>
</div>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
    test(() => {
        // Make the content unfragmented. That should clear all pagination
        // struts that were needed while we were fragmented.
        document.body.offsetTop;
        document.getElementById("multicol").style.columns = "auto";

        assert_equals(document.getElementById("inner").offsetHeight, 140);
    }, "No pagination struts should be left behind.");
</script>
